RF: more of refactoring between http and s3
